refactor(docs): split docs-map render tests by behavior
